agent/pyproject.toml
[project]
name = "stirr-a2ui-agent"
version = "0.1.0"
description = "STIRR A2A agent for ConversationalSearch + ContentShelf. Gemini-powered."
readme = "../README.md"
requires-python = ">=3.13"
dependencies = [
    "a2a-sdk>=0.3.0",
    "httpx>=0.27.0",
    "uvicorn[standard]>=0.30.0",
    "starlette>=0.38.0",
    "click>=8.1.8",
    "google-adk>=1.8.0",
    "google-genai>=1.27.0",
    "python-dotenv>=1.1.0",
    "litellm",
    "jsonschema>=4.0.0",
    "a2ui-agent",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
